Move more scene tests to the right folder
